refactor: simplified button component
