+++ plugins/login-plugin/src/Components/UIComponents/Buttons/FocusableButton.js
@@ -25,46 +25,60 @@ const localStyles = StyleSheet.create({
   },
 });
 
-const Button = ({
-  onPress,
-  label,
-  groupId,
-  backgroundColor,
-  backgroundColorFocused,
-  textColorFocused,
-  textStyles,
-  borderRadius,
-}) => {
-  const getBackgroundColor = React.useCallback(
-    (focused) => ({
-      backgroundColor: focused ? backgroundColorFocused : backgroundColor,
-    }),
-    []
-  );
+const Button = React.forwardRef(
+  (
+    {
+      onPress,
+      label,
+      groupId,
+      backgroundColor,
+      backgroundColorFocused,
+      textColorFocused,
+      textStyles,
+      borderRadius,
+      nextFocusUp,
+      nextFocusDown,
+    },
+    ref
+  ) => {
+    const getBackgroundColor = React.useCallback(
+      (focused) => ({
+        backgroundColor: focused ? backgroundColorFocused : backgroundColor,
+      }),
+      []
+    );
 
-  const getTextColor = React.useCallback(
-    (focused) => ({
-      color: focused ? textColorFocused : textStyles.color,
-    }),
-    []
-  );
+    const getTextColor = React.useCallback(
+      (focused) => ({
+        color: focused ? textColorFocused : textStyles.color,
+      }),
+      []
+    );
 
-  return (
-    <Focusable id={`${groupId}-${label}`} groupId={groupId} onPress={onPress}>
-      {(focused) => (
-        <View
-          style={[
-            localStyles.button,
-            getBackgroundColor(focused),
-            { borderRadius },
-          ]}
-        >
-          <Text style={[textStyles, getTextColor(focused)]}>{label}</Text>
-        </View>
-      )}
-    </Focusable>
-  );
-};
+    return (
+      <Focusable
+        ref={ref}
+        id={`${groupId}-${label}`}
+        groupId={groupId}
+        onPress={onPress}
+        nextFocusUp={nextFocusUp}
+        nextFocusDown={nextFocusDown}
+      >
+        {(focused) => (
+          <View
+            style={[
+              localStyles.button,
+              getBackgroundColor(focused),
+              { borderRadius },
+            ]}
+          >
+            <Text style={[textStyles, getTextColor(focused)]}>{label}</Text>
+          </View>
+        )}
+      </Focusable>
+    );
+  }
+);
 
 Button.propTypes = {
   onPress: PropTypes.func,
@@ -76,6 +90,8 @@ Button.propTypes = {
   textColorFocused: PropTypes.string,
   textStyles: PropTypes.object,
   borderRadius: PropTypes.number,
+  nextFocusUp: PropTypes.object,
+  nextFocusDown: PropTypes.object,
 };
 
 Button.defaultProps = {
@@ -87,5 +103,7 @@ Button.defaultProps = {
   textStyles: {},
   borderRadius: 5,
 };
+
+Button.displayName = "Button";
 
 export default Button;
